Extracts interface for RemotePipeline to allow other implementations besides HTTP
