+++ api/src/main/java/com/findwise/hydra/local/HttpEndpointConstants.java
@@ -1,0 +1,20 @@
+package com.findwise.hydra.local;
+
+public class HttpEndpointConstants {
+    public static final String GET_DOCUMENT_URL = "getDocument";
+    public static final String WRITE_DOCUMENT_URL = "writeDocument";
+    public static final String RELEASE_DOCUMENT_URL = "releaseDocument";
+    public static final String PROCESSED_DOCUMENT_URL = "processedDocument";
+    public static final String PENDING_DOCUMENT_URL = "pendingDocument";
+    public static final String DISCARDED_DOCUMENT_URL = "discardedDocument";
+    public static final String GET_PROPERTIES_URL = "getProperties";
+    public static final String FAILED_DOCUMENT_URL = "failedDocument";
+    public static final String FILE_URL = "documentFile";
+    public static final String STAGE_PARAM = "stage";
+    public static final String NORELEASE_PARAM = "norelease";
+    public static final String PARTIAL_PARAM = "partial";
+    public static final String DOCID_PARAM = "docid";
+    public static final String FILENAME_PARAM = "filename";
+    public static final int DEFAULT_PORT = 12001;
+    public static final String DEFAULT_HOST = "localhost";
+}
